requirements.txt
# Core dependencies
python-dotenv>=1.0.0
PyYAML>=6.0
cryptography>=41.0.0
requests>=2.31.0
APScheduler>=3.10.0
SQLAlchemy>=2.0.0
Flask>=3.0.0
Flask-SocketIO>=5.0.0
python-socketio[client]>=5.0.0

# Logging
structlog>=23.0.0

# Testing (dev)
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Type checking (dev)
mypy>=1.0.0
